requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.5.1
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
compress_json==1.1.0
contourpy==1.2.1
cupy-cuda12x==13.3.0
cycler==0.12.1
Cython==3.0.11
databricks-sdk==0.29.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dm-tree==0.1.8
docker==7.1.0
docstring_parser==0.16
dynaconf==3.2.6
entrypoints==0.4
etils==1.9.2
executing==2.0.1
Faker==28.1.0
fastapi==0.112.1
fastjsonschema==2.20.0
fastrlock==0.8.2
ffmpy==0.4.0
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
funcy==2.0
gast==0.6.0
gensim==4.3.3
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
gradio==4.41.0
gradio_client==1.3.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.65.4
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
immutabledict==4.2.0
importlib_metadata==7.2.1
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
iterative-telemetry==0.0.8
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.6.17
keras==3.5.0
keras-tuner==1.4.7
kiwisolver==1.4.5
kt-legacy==1.0.5
langdetect==1.0.9
libclang==18.1.1
litestar==2.11.0
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mlflow==2.15.1
mlflow-skinny==2.15.1
msgspec==0.18.6
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.8.1
notebook_shim==0.2.4
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
oauth2client==4.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
overrides==7.7.0
p2j==1.3.2
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plot_keras_history==1.1.36
plotly==5.24.0
pluggy==1.5.0
polyfactory==2.16.2
portalocker==2.10.1
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.1
pydub==0.25.1
Pygments==2.18.0
pyLDAvis==3.4.1
pyparsing==3.1.2
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
querystring-parser==1.2.4
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rich-click==1.8.3
rpds-py==0.19.1
rsa==4.9
ruff==0.6.0
sacrebleu==2.4.2
safetensors==0.4.4
sanitize_ml_labels==1.1.0
scikit-learn==1.5.1
scikit-multilearn==0.2.0
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
seqeval==1.2.2
setuptools==71.1.0
shellingham==1.5.4
simple_parsing==0.1.5
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
sqlparse==0.5.1
stack-data==0.6.3
StackAPI==0.3.1
starlette==0.38.2
statsmodels==0.14.2
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-datasets==4.9.6
tensorflow-hub==0.16.1
tensorflow-metadata==1.15.0
tensorflow-model-optimization==0.8.0
tensorrt==10.3.0
tensorrt-cu12==10.3.0
tensorrt-cu12-bindings==10.3.0
tensorrt-cu12-libs==10.3.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tf-models-no-deps==2.17.0
tf-slim==1.1.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchdog==5.0.0
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.44.0
wordcloud==1.9.3
wrapt==1.16.0
xgboost==2.1.1
zipp==3.19.2
